Added tests for SubmitActionViewModel, also added some shanges in SubmitActionViewModel
